refactor(node): Extract `Yarn2Command` to a dedicated file
Improve the overview a bit.

While at it, also move `isCorepackEnabledInManifest`, as `Yarn2Command`
is the only caller.
